utils: Move more repository functionality into repo-utils
