+++ app/flatpak-builtins-build-bundle.c
@@ -34,6 +34,7 @@
 #include "libglnx.h"
 
 #include "flatpak-builtins.h"
+#include "flatpak-repo-utils-private.h"
 #include "flatpak-utils-private.h"
 #include "flatpak-oci-registry-private.h"
 #include "flatpak-chain-input-stream-private.h"
