+++ po/POTFILES.in
@@ -59,6 +59,7 @@ common/flatpak-oci-registry.c
 common/flatpak-progress.c
 common/flatpak-ref-utils.c
 common/flatpak-remote.c
+common/flatpak-repo-utils.c
 common/flatpak-run.c
 common/flatpak-run-dbus.c
 common/flatpak-transaction.c
